Passing all valid rv32-i tests for the assembler. Run `llvm-lit -v MERRY` in the test/MC directory to see results.
Some things to note: I have commented out most assembler aliases. These
do not impact correctness but simply emit code that is easier to read.
There is no easy way to translate X0 operands in tablegen to DC0
operands. Aliases would have to be coded in C++ which is not worthwhile
at the moment.

I had to comment out some patterns for selection and for pseudo
instructions. These must be added in C++ for proper code generation, but
for the purposes of the assembler we don't need them.

For printing DC operands, I used getImm to get the offset. This is
something to look into for later, I might need to add some sort of DC
type to higher levels of the compiler for code generation.